refactor(ui): uniformise les noms de fichiers et composants, ajuste les imports
- Renomme les fichiers et classes pour une meilleure cohérence (`ToastComponent` → `Toast`, `ToastContainerComponent` → `ToastContainer`).
- Réorganise les fichiers dans des dossiers appropriés (`login-form`, `login-modal`, `toast`, etc.).
- Met à jour les imports globaux dans `index.ts` et les composants concernés (`App`, `ToastContainer`, `LoginModal`, etc.).
- Corrige les chemins et alias pour refléter les changements dans la structure du projet.
