+++ src/app/shared/ui/index.ts
@@ -1,13 +1,13 @@
 // Export all UI components for easier imports
 export { NavbarComponent } from './navbar/navbar';
-export { LoginForm } from './login-form/login-form';
-export { LoginModal } from './login-modal/login-modal';
-export { ToastComponent } from './toast/toast.component';
-export { ToastContainerComponent } from './toast/toast-container.component';
+export { LoginForm } from '@shared/ui/login/login-form/login-form';
+export { LoginModal } from '@shared/ui/login/login-modal/login-modal';
+export { Toast } from './toast/toast';
+export { ToastContainer } from './toast/toast-container';
 
 // Export services
 export { AuthService } from '../../core/services/auth.service';
-export { ToastService } from './toast/toast.service';
+export { ToastService } from './toast/service/toast-service';
 
 // Export types
 export type { LoginRequest, User, AuthState } from '../../core/services/auth.service';
